change styling for login screena and refactored the code a lil bit
