+++ android/app/build.gradle
@@ -33,7 +33,7 @@ android {
 
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
-        applicationId "relevantsystems.relevantsystems.com.relevant_systems"
+        applicationId "com.relevantSystems.relevant_systems"
         minSdkVersion 16
         targetSdkVersion 28
         versionCode flutterVersionCode.toInteger()
